Changed the endpoint to /auth
